separate TwoButtonControl to separate file
